Fix bug preventing multiple files from being selected
- Steps to reproduce (heal preprod):
  - Search for 'jcoin' and select that study
  - See 1 study selected
  - Search for 'bacpac' and select that study
  - Expect 2 studies selected, get 1 study selected

- Cause: table selection onChange handler sets selected resources to be
whatever's currently selected in the table. If search or filtering has removed
a selected row in the table, the onChange handler will happily drop it, as
even though it was selected it's no longer in the table.

- Fix: ~~Update table selection onChange handler to work by deselecting / selecting
a single row at a time.~~ Actually it turns out there's a one-line config change
that also fixes the issue :angry-haha: