fix: disable query retries & align storybook request config w/ real app
For some reason, storybook was using a different config for react-query than the app,
leading to confusion and impossible to reproduce/mock scenarios.
This commit also makes sure that webservice errors are explicitly catched and thrown. The
previous code only worked for errors that happen to not have a json text in them.
Finally, this commit disables retries of failed requests at a global level for the GWAS app.
See https://react-query-v3.tanstack.com/guides/query-retries